# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Payment Platform API Gateway
  description: Public HTTP API exposed by the api-gateway service.
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://api.payment-platform.local
    description: Kubernetes ingress

tags:
  - name: auth
    description: Authentication and token management
  - name: payments
    description: Payment lifecycle
  - name: wallet
    description: User wallet balances
  - name: refunds
    description: Payment refunds
  - name: admin
    description: Administrative operations (requires admin role)

paths:
  /health:
    get:
      summary: Gateway health check
      operationId: getHealth
      tags: [auth]
      security: []
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok

  /auth/signup:
    post:
      summary: Register a new user
      operationId: signup
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"

  /auth/login:
    post:
      summary: Authenticate and obtain tokens
      operationId: login
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /auth/refresh:
    post:
      summary: Refresh access token
      operationId: refreshToken
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: New tokens issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /payments:
    post:
      summary: Create a payment
      operationId: createPayment
      tags: [payments]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentRequest"
      responses:
        "201":
          description: Payment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"

  /payments/{id}:
    get:
      summary: Get payment by ID
      operationId: getPayment
      tags: [payments]
      parameters:
        - $ref: "#/components/parameters/PaymentId"
      responses:
        "200":
          description: Payment details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  /refunds:
    post:
      summary: Refund a payment
      operationId: refundPayment
      tags: [refunds]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefundRequest"
      responses:
        "200":
          description: Refund processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  /wallet:
    get:
      summary: Get wallet balance for authenticated user
      operationId: getWallet
      tags: [wallet]
      responses:
        "200":
          description: Wallet balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /admin/payments:
    get:
      summary: List payments (admin)
      operationId: adminListPayments
      tags: [admin]
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, completed, failed, refunded]
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: Paginated payment list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

  /admin/payments/{id}:
    get:
      summary: Get payment details (admin)
      operationId: adminGetPayment
      tags: [admin]
      parameters:
        - $ref: "#/components/parameters/PaymentId"
      responses:
        "200":
          description: Payment details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  /admin/audit-events:
    get:
      summary: List audit events (admin)
      operationId: adminListAuditEvents
      tags: [admin]
      parameters:
        - name: resource_type
          in: query
          schema:
            type: string
        - name: resource_id
          in: query
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        "200":
          description: Audit event list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditEventListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

  /admin/payments/{id}/force-refund:
    post:
      summary: Force refund a payment (admin)
      operationId: adminForceRefund
      tags: [admin]
      parameters:
        - $ref: "#/components/parameters/PaymentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminForceRefundRequest"
      responses:
        "200":
          description: Refund initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    PaymentId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      description: Unique key for idempotent payment creation
      schema:
        type: string
        maxLength: 128

  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    RateLimited:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

  schemas:
    SignupRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string

    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer
          format: int64
        user_id:
          type: string
          format: uuid
        email:
          type: string
          format: email

    CreatePaymentRequest:
      type: object
      required: [amount, currency]
      properties:
        amount:
          type: number
          format: double
          exclusiveMinimum: 0
        currency:
          type: string
          example: USD
        provider:
          type: string
          example: stripe

    RefundRequest:
      type: object
      required: [payment_id, amount]
      properties:
        payment_id:
          type: string
          format: uuid
        amount:
          type: number
          format: double
          exclusiveMinimum: 0

    AdminForceRefundRequest:
      type: object
      required: [amount, reason]
      properties:
        amount:
          type: number
          format: double
          exclusiveMinimum: 0
        reason:
          type: string
          maxLength: 500

    PaymentResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        amount:
          type: number
          format: double
        currency:
          type: string
        status:
          type: string
          enum: [pending, completed, failed, refunded]
        provider:
          type: string

    PaymentListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/PaymentResponse"
        total:
          type: integer

    WalletResponse:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
        balance:
          type: number
          format: double
        version:
          type: integer

    AuditEvent:
      type: object
      properties:
        id:
          type: string
          format: uuid
        actor_id:
          type: string
          format: uuid
        action:
          type: string
        resource_type:
          type: string
        resource_id:
          type: string
        created_at:
          type: string
          format: date-time

    AuditEventListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AuditEvent"

    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        code:
          type: string

security:
  - bearerAuth: []